Step6 PR (#58)
* 답변 삭제 구현

* 답변 개수 데이터 반영하기

* swagger기반 테스트 적용

+++ src/main/java/codesquad/web/ApiAnswerController.java
@@ -3,12 +3,12 @@ package codesquad.web;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import codesquad.domain.Answer;
 import codesquad.domain.AnswerRepository;
@@ -17,9 +17,9 @@ import codesquad.domain.QuestionRepository;
 import codesquad.domain.Result;
 import codesquad.domain.User;
 
-@Controller
-@RequestMapping("/questions/{questionId}/answers")
-public class AnswerController {
+@RestController
+@RequestMapping("/api/questions/{questionId}/answers")
+public class ApiAnswerController {
 	@Autowired
 	private QuestionRepository questionRepository;
 	
@@ -27,26 +27,31 @@ public class AnswerController {
 	private AnswerRepository answerRepository;
 	
 	@PostMapping("")
-	public String create(@PathVariable Long questionId, String contents, HttpSession session) {
+	public Answer create(@PathVariable Long questionId, String contents, HttpSession session) {
 		if (!HttpSessionUtils.isLoginUser(session)) {
-			return "redirect:/users/loginForm";
+			return null;
 		}
 		User loginUser = (User) HttpSessionUtils.getUserFromSession(session);
 		Question question = questionRepository.findById(questionId).get();
 		Answer answer = new Answer(loginUser, question, contents);
-		answerRepository.save(answer);
-		return String.format("redirect:/questions/%d", questionId);
+		question.addAnswer();
+		return answerRepository.save(answer);
 	}
 	
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long questionId, @PathVariable Long id, Model model, HttpSession session) {
+	public Result delete(@PathVariable Long questionId, @PathVariable Long id, HttpSession session) {
+		if (!HttpSessionUtils.isLoginUser(session)) {
+			return Result.fail("로그인 해야 합니다");
+		}
 		Answer answer = answerRepository.findById(id).get();
 		Result result = Result.valid(session, answer.classForQAndA());
 		if (!result.isValid()) {
-			model.addAttribute("errorMessage", result.getErrorMessage());
-			return "/user/login_failed";
+			return Result.fail("자신의 글만 삭제할 수 있습니다!!!");
 		}
 		answerRepository.delete(answer);
-		return String.format("redirect:/questions/%d", questionId);
+		Question question = questionRepository.findById(questionId).get();
+		question.delteAnswer();
+		questionRepository.save(question);
+		return Result.ok();
 	}
 }
